+++ src/Entities/BankPaymentResponse.php
@@ -11,48 +11,24 @@ class BankPaymentResponse
 {
     /**
      * The open banking transaction id
-     * @var string
      */
-    public $id;
-
+    public string $id;
     /**
      * URL to redirect the customer to
      * Sent there so merchant can redirect consumer to complete the payment.
-     *
-     * @var string
      */
-    public $redirectUrl;
-
-    /** @var string */
-    public $paymentStatus;
-
+    public ?string $redirectUrl;
+    public ?string $paymentStatus;
     /** @var BankPaymentType */
     public $type;
-
-    /** @var string */
-    public $tokenRequestId;
-
-    /** @var string */
-    public $sortCode;
-
-    /** @var string */
-    public $accountName;
-
-    /** @var string */
-    public $accountNumber;
-
-    /** @var string */
-    public $iban;
-
-    /** @var string */
-    public $remittanceReferenceValue;
-
-    /** @var string */
-    public $remittanceReferenceType;
-
-    /** @var float */
-    public $amount;
-
-    /** @var string */
-    public $currency;
+    public ?string $tokenRequestId;
+    public ?string $sortCode;
+    public ?string $accountName;
+    public ?string $accountNumber;
+    public ?string $iban;
+    public ?string $remittanceReferenceValue;
+    public ?string $remittanceReferenceType;
+    public ?float $amount;
+    public ?string $currency;
+    public ?string $maskedIbanLast4;
 }